update multihumandata & unitest
